Move error reporting into LoxError Module
