+++ src/scanner.rs
@@ -1,6 +1,5 @@
 use crate::{
     error::{LoxError, LoxErrorKind},
-    lox::Lox,
     token::*,
     value::Value,
 };
@@ -76,7 +75,7 @@ impl Scanner {
         }
 
         if self.is_at_end() {
-            return Err(Lox::error(
+            return Err(LoxError::error(
                 self.line,
                 "Unterminated string",
                 LoxErrorKind::SyntaxError,
@@ -273,7 +272,7 @@ impl Scanner {
             }
 
             _ => {
-                return Err(Lox::error(
+                return Err(LoxError::error(
                     line,
                     "Unexpected character",
                     LoxErrorKind::SyntaxError,
